Rename various query utility classes to use 'spawndatastore' tag instead of 'zookeeper'
