+++ hydra-main/src/test/java/com/addthis/hydra/query/QueryConfigWatcherTest.java
@@ -22,7 +22,7 @@ import com.addthis.hydra.job.JobQueryConfig;
 import com.addthis.hydra.job.ZnodeJob;
 import com.addthis.hydra.job.store.SpawnDataStore;
 import com.addthis.hydra.job.store.ZookeeperDataStore;
-import com.addthis.hydra.query.zookeeper.QueryConfigWatcher;
+import com.addthis.hydra.query.spawndatastore.QueryConfigWatcher;
 
 import org.junit.Before;
 import org.junit.Test;
